scanner-backend/requirements.txt
opencv-python>=4.8.0
pyzbar>=0.1.9
Pillow>=10.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
websockets>=12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
requests>=2.31.0
